Assignment probs based on list sizes
Patch to assign members to each list with a probability based on the relative size of each list to generate. This fixes a bug in which members were assigned to lists with equal probability (until each list hit is specified size), causing small lists to get filled up with older members (lower primary keys) and larger lists to over-represent newer members.